requirements.txt
torch>=1.10.0
transformers>=4.20.0
tree-sitter>=0.20.0
tree-sitter-java>=0.20.0
numpy>=1.21.0
scikit-learn>=1.0.0
networkx>=2.6.0
pandas>=1.3.0
chromadb>=0.3.0
sentence-transformers>=2.2.0
javalang>=0.13.0
pathlib2>=2.3.0
tqdm>=4.62.0
json5>=0.9.0
faiss-cpu>=1.7.0 